+++ dogfooding/lib/di/injector.dart
@@ -13,6 +13,7 @@ import 'package:stream_video_noise_cancellation/noise_cancellation_audio_process
 import 'package:stream_video_push_notification/stream_video_push_notification.dart';
 
 import '../app/user_auth_controller.dart';
+import '../core/repos/custom_environment_loader.dart';
 import '../core/repos/token_service.dart';
 import '../core/repos/user_auth_repository.dart';
 import '../log_config.dart';
@@ -24,7 +25,9 @@ GetIt locator = GetIt.instance;
 /// and injecting them into the app.
 class AppInjector {
   // Register dependencies
-  static Future<void> init({Environment? forceEnvironment}) async {
+  static Future<void> init({
+    Environment? forceEnvironment,
+  }) async {
     // Google sign in
     locator.registerSingleton<GoogleSignIn>(
       GoogleSignIn(hostedDomain: 'getstream.io'),
@@ -41,7 +44,11 @@ class AppInjector {
     locator.registerSingleton<AppPreferences>(appPrefs);
 
     // Repositories
-    locator.registerSingleton(const TokenService());
+    locator.registerSingleton(
+      TokenService(
+        customEnvironmentLoader: customEnvironmentLoader,
+      ),
+    );
 
     locator.registerFactoryParam<UserAuthRepository, User, TokenResponse>(
       (user, tokenResponse) {