+++ packages/stream_video/lib/src/coordinator/open_api/coordinator_client_open_api.dart
@@ -442,6 +442,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     bool? notify,
     bool? video,
     DateTime? startsAt,
+    int? membersLimit,
     CallSettingsRequest? settingsOverride,
     Map<String, Object> custom = const {},
   }) async {
@@ -469,6 +470,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
             settingsOverride: settingsOverride,
             custom: custom,
           ),
+          membersLimit: membersLimit,
           ring: ringing,
           notify: notify,
           video: video,
@@ -507,6 +509,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
     bool? create,
     String? migratingFrom,
     bool? video,
+    int? membersLimit,
   }) async {
     try {
       _logger.d(
@@ -528,6 +531,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
           create: create,
           ring: ringing,
           location: location,
+          membersLimit: membersLimit,
           migratingFrom: migratingFrom,
           video: video,
         ),
@@ -677,6 +681,64 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
       );
       if (result == null) {
         return Result.error('requestPermissions result is null');
+      }
+      return const Result.success(none);
+    } catch (e, stk) {
+      return Result.failure(VideoErrors.compose(e, stk));
+    }
+  }
+
+  @override
+  Future<Result<None>> videoPin({
+    required StreamCallCid callCid,
+    required String sessionId,
+    required String userId,
+  }) async {
+    try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[videoPin] no connection established');
+        return connectionResult;
+      }
+      final result = await _defaultApi.videoPin(
+        callCid.type.value,
+        callCid.id,
+        open.PinRequest(
+          sessionId: sessionId,
+          userId: userId,
+        ),
+      );
+      if (result == null) {
+        return Result.error('[videoPin] result is null');
+      }
+      return const Result.success(none);
+    } catch (e, stk) {
+      return Result.failure(VideoErrors.compose(e, stk));
+    }
+  }
+
+  @override
+  Future<Result<None>> videoUnpin({
+    required StreamCallCid callCid,
+    required String sessionId,
+    required String userId,
+  }) async {
+    try {
+      final connectionResult = await _waitUntilConnected();
+      if (connectionResult is Failure) {
+        _logger.e(() => '[videoUnpin] no connection established');
+        return connectionResult;
+      }
+      final result = await _defaultApi.videoUnpin(
+        callCid.type.value,
+        callCid.id,
+        open.UnpinRequest(
+          sessionId: sessionId,
+          userId: userId,
+        ),
+      );
+      if (result == null) {
+        return Result.error('[videoUnpin] result is null');
       }
       return const Result.success(none);
     } catch (e, stk) {
@@ -973,7 +1035,7 @@ class CoordinatorClientOpenApi extends CoordinatorClient {
   @override
   Future<Result<QueriedMembers>> queryMembers({
     required StreamCallCid callCid,
-    required Map<String, Object> filterConditions,
+    Map<String, Object> filterConditions = const {},
     String? next,
     String? prev,
     List<open.SortParamRequest> sorts = const [],